3AC for Methods and Object

+++ tests/test_14.java
@@ -1,5 +1,5 @@
 public class hello{
-    int fun(int arr[], int b, int c){
+    int fun(double arr[], int b, int c){
         int k= arr[b];
         String str;
         str = "abc" + b + c;
@@ -8,7 +8,7 @@ public class hello{
 
     void main(){
         int a,b,c,d;
-        float arr[][] = {{1,2,3},{4, 5, 6}};
+        double arr[][] = {{1.0,2.0,3.0},{4.0, 5.0, 6.0}};
         b=1;
         c=1;
         int e = fun(arr[0], b, c);
